chore: 부적절한 위치에 있는 테스트 코드 위치 조정
